# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portal Berita",
        "description": "Portal Berita API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local server"
        },
        {
            "url": "http://staging.example.com",
            "description": "Staging server"
        },
        {
            "url": "http://example.com",
            "description": "Production server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/transaksi-masuk": {
            "get": {
                "tags": [
                    "Transaksi Masuk"
                ],
                "summary": "List of Transaksi Masuk",
                "description": "Retrieve a list of transaksi masuk",
                "operationId": "listTransaksiMasuk",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Successfully retrieved transaksi masuk",
                                    "data": [
                                        {
                                            "id": 1,
                                            "barang_id": 2,
                                            "supplier_id": 1,
                                            "tanggal": "2024-04-30",
                                            "jumlah": 5,
                                            "harga_beli": 100000
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transaksi Masuk"
                ],
                "summary": "Create a new transaksi masuk",
                "description": "Add a new transaksi masuk record",
                "operationId": "createTransaksiMasuk",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "barang_id",
                                    "supplier_id",
                                    "tanggal",
                                    "jumlah",
                                    "harga_beli"
                                ],
                                "properties": {
                                    "barang_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-30"
                                    },
                                    "jumlah": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "harga_beli": {
                                        "type": "number",
                                        "example": 100000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaksi masuk created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Transaksi Masuk created successfully",
                                    "data": {
                                        "id": 1,
                                        "barang_id": 2,
                                        "supplier_id": 1,
                                        "tanggal": "2024-04-30",
                                        "jumlah": 5,
                                        "harga_beli": 100000
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transaksi-masuk/search": {
            "get": {
                "tags": [
                    "Transaksi Masuk"
                ],
                "summary": "Search transaksi masuk by date",
                "description": "Search transaksi masuk records by exact date",
                "operationId": "searchTransaksiMasukByTanggal",
                "parameters": [
                    {
                        "name": "tanggal",
                        "in": "query",
                        "description": "Tanggal transaksi dalam format YYYY-MM-DD",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-04-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matching transaksi masuk found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Transaksi Masuk found",
                                    "data": [
                                        {
                                            "id": 1,
                                            "barang_id": 2,
                                            "supplier_id": 1,
                                            "tanggal": "2024-04-30",
                                            "jumlah": 5,
                                            "harga_beli": 100000
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No matching transaksi masuk found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "No transaksi masuk found on this date"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transaksi-masuk/{id}": {
            "get": {
                "tags": [
                    "Transaksi Masuk"
                ],
                "summary": "Get transaksi masuk by ID",
                "description": "Retrieve a single transaksi masuk record by its ID",
                "operationId": "getTransaksiMasukById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaksi Masuk found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Transaksi Masuk found",
                                    "data": {
                                        "id": 1,
                                        "barang_id": 2,
                                        "supplier_id": 1,
                                        "tanggal": "2024-04-30",
                                        "jumlah": 5,
                                        "harga_beli": 100000
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaksi Masuk not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Transaksi Masuk not found"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Transaksi Masuk"
                ],
                "summary": "Update a transaksi masuk",
                "description": "Update an existing transaksi masuk",
                "operationId": "updateTransaksiMasuk",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "barang_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-01"
                                    },
                                    "jumlah": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "harga_beli": {
                                        "type": "number",
                                        "example": 120000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaksi Masuk updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Transaksi Masuk updated successfully",
                                    "data": {
                                        "id": 1,
                                        "barang_id": 3,
                                        "supplier_id": 2,
                                        "tanggal": "2024-05-01",
                                        "jumlah": 10,
                                        "harga_beli": 120000
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transaksi Masuk"
                ],
                "summary": "Delete a transaksi Masuk",
                "description": "Delete a transaksi masuk record by ID",
                "operationId": "deleteTransaksiMasuk",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaksi Masuk deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Transaksi Masuk deleted successfully"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}